+++ lib/main.dart
@@ -1,8 +1,7 @@
-import 'dart:developer';
-
 import 'package:flutter/material.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_local_notifications/flutter_local_notifications.dart';
+import 'package:flutter_logs/flutter_logs.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:hive/hive.dart';
 import 'package:hive_flutter/hive_flutter.dart';
@@ -74,7 +73,7 @@ void main() async {
   await flutterLocalNotificationsPlugin.initialize(initializationSettings,
       onSelectNotification: (String? payload) async {
     if (payload != null) {
-      log('notification payload: $payload');
+      FlutterLogs.logInfo('notification', 'payload', payload);
     }
   });
 
@@ -95,6 +94,25 @@ void main() async {
       walletToOpenDirectly: notificationAppLaunchDetails?.payload ?? '',
     );
   }
+
+  //init logger
+  await FlutterLogs.initLogs(
+    logLevelsEnabled: [
+      LogLevel.INFO,
+      LogLevel.WARNING,
+      LogLevel.ERROR,
+      LogLevel.SEVERE
+    ],
+    timeStampFormat: TimeStampFormat.TIME_FORMAT_READABLE,
+    directoryStructure: DirectoryStructure.FOR_DATE,
+    logFileExtension: LogFileExtension.LOG,
+    logsWriteDirectoryName: 'MyLogs',
+    logsExportDirectoryName: 'MyLogs/Exported',
+    debugFileOperations: true,
+    isDebuggable: true,
+  );
+
+  FlutterLogs.logInfo('main', 'initLogs', 'Init logs..');
 
   //run
   runApp(PeercoinApp());